requirements_wgan_diffusion.txt
numpy==1.24.3
pandas==2.0.3
tensorflow==2.13.0
scikit-learn==1.3.2
matplotlib==3.7.5
umap-learn==0.5.7
keras==2.13.1
statsmodels==0.14.1
torch==2.4.1
seaborn==0.12.2